Esquema
Cambio de esquema y nuevo algoritmo para ROSA
