+++ ergor/test/Metodo Rosa.txt
